heavily refactoring the electron app
