+++ .vscode/settings.json
@@ -1,4 +1,9 @@
 {
     "python.testing.unittestEnabled": false,
-    "python.testing.pytestEnabled": true
+    "python.testing.pytestEnabled": true,
+    "editor.formatOnSave": true,
+    "editor.defaultFormatter": "esbenp.prettier-vscode",
+    "[javascript]": {
+       "editor.defaultFormatter": "esbenp.prettier-vscode"
+    },
 }
